add error message

+++ backend/apps/main/serializers/ftx.py
@@ -2,8 +2,8 @@ from rest_framework import serializers
 
 
 class FTXOrderPlaceSerializer(serializers.Serializer):
-    quantity = serializers.DecimalField(max_digits=10, decimal_places=10)
-    price = serializers.IntegerField()
+    quantity = serializers.DecimalField(max_digits=20, decimal_places=10)
+    price = serializers.DecimalField(max_digits=20, decimal_places=10)
     symbol = serializers.CharField()
     trade_type = serializers.CharField()
     bot_type = serializers.CharField()
